Update patch set 3
Patch Set 3: -Presubmit-Ready

Thanks for the feedback and review. I am glad for the opportunity to use a CL as a means to discuss a point and come to a conclusion. I appreciate the time you granted towards this.

After all ideas have been discussed and hashed out, ultimately it is the responsibility of the developer to heed or ignore the messaging. Erring on the side of caution for the messaging is for the best.

I have decided to abandon this and close the bug(s) as the current messaging is the new normal.

Patch-set: 3
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0