Update patch set 12
Patch Set 12:

(1 comment)

FWIW I am retesting the boot impact (it was done when I _first_ composed this CL, and not recently after all the structural changes in init etc) and the average response time. As well as an if (getpid() == 1) return false to remove the impact if any.

Patch-set: 12
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>