Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> +victoryang
> 
> this seems like something the Go folks will need to override?

Thanks for bringing this to our attention! Indeed this is something that we'd like to test and possibly override. I've cc'ed the team in the bug and it looks like there's already a request for perf test in the bug, so we can continue the discussion there.

Patch-set: 1
Reviewer: Gerrit User 1349653 <1349653@85c56323-6fa9-3386-8a01-6480fb634889>