init: first_stage_init parse /init.rc
First stage init provides a means to mount system and vendor
partitions.  Some hardware requires kernel modules to be loaded
in order to access the related storage.  To address this, parse
/init.rc and push a first-stage-init trigger.

A Proposed means to build and initialize the ramdisk with the
necessary pieces in the device specific configuration to load a
boot module follows:

file init.first_stage_init.rc:

+on first-stage-init
+    insmod [-f] <module>.ko [<arguments>]

file device.mk:

+PRODUCT_COPY_FILES += \
+    device/google/<device>/init.first_stage_init.rc:$(TARGET_COPY_OUT_RECOVERY)/root/first_stage_ramdisk/init.rc \
+    device/google/<device>-kernel/<module>.ko:$(TARGET_COPY_OUT_RECOVERY)/root/first_stage_ramdisk/<module>.ko

Hardened insmod handler because syntax mistakes here do not land in
convenient logs and result in tighter boot loops.

Audit set of commands that make no sense in first stage init, and if
operated, report an error.

Test: boot, inspect log messages for contrived configuration.
Test: recovery, confirm no parser complaints about its /init.rc
Bug: 129780532
Change-Id: Ie0700f187d9516bd1a52e3e71a17f01b337458f8