zip_archive.cc: fix ubsan false positives
std::hash returns a 64 bit value, which is truncated to a 32 bit value
in ComputeHash. ubsan's implicit-unsigned-integer-truncation doesn't
like this implicit truncation and crashes the program. Explicitly strip
off the top order bits after computing the hash.

Set __attribute__((no_sanitize("integer"))) on the entire function. Hash
computation inherently relies on unusual integer operations which have
no security implications. Setting this attribute eliminates the need
for the truncation described above, but we preserve it anyway for
clarity.

Setting the attribute is also needed for the WIN32 version of the
hashing code.

This is needed to support Android booting with ubsan's
implicit-unsigned-integer-truncation option enabled.

Test: compiles and boots
Bug: 122975762
Change-Id: I2f05fbf5ffee8e90a66a6fda32e80de9cca246c0