libziparchive: clean up some -Wconversion warnings
The third argument of the crc32() function is an unsigned int, not a
size_t. As a result, truncation occurs when calling this function,
making it 64 bit unsafe. Use crc32_z() instead of crc32(), where the
third argument is a size_t. See https://www.zlib.net/manual.html#Checksum

Contrary to their names, both crc32() and crc32_z() return an "unsigned
long", which is not necessarily a 32 bit value. According to Mark
Adler (zlib maintainer), the choice of unsigned long was made 24 years
ago, prior to the existence of fixed width types like uint32_t. For
compatibility reasons, the return value cannot be changed. However, the
functions effectively guarantee a return value in the range of 0 to
UINT32_MAX, so we can safely cast to the smaller type without the
potential for data loss. While it may be obvious from the function name,
Mark will be updating the zlib documentation to be clearer that the
return value is constrained to 32 bit numbers.

Test: compiles and boots
Test: atest ziparchive-tests
Change-Id: I3803a039771199093a2e82c354b987028cc621c4