Update patch set 1
Patch Set 1:

(1 comment)

> Patch Set 1:
> 
> I'd like a shell settable system property (e.g. sys.audit.rate_limit) that can be set to change or remove the rate limit. Could that same property be used here? i.e. have init add the rate limit on the very last init trigger?

The downside of this approach is that I would have to query the rate_limit property every time an SELinux denial is generated, bypassing an optimization I put in place to avoid this. I guess that's not that bad. Let me look into it...

Patch-set: 1
Reviewer: Gerrit User 1003966 <1003966@85c56323-6fa9-3386-8a01-6480fb634889>