Update patch set 3
Patch Set 3: Code-Review+2

maybe we should consider a mid tier setting for things in TA and backound policy, but that's kind of late, also we can consider put Java API for app in this group. Anyway, it is fine at this moment to make things in TA as before blkio.

Patch-set: 3
Reviewer: Gerrit User 1105075 <1105075@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2