bootstat: Allow regex matches for boot reasons
If the matching string contains a [.], or a \., then it is a regex
match.  If we do not find an exact string match, switch to iterate
through the entire list for regex strings to find a match.

This allows us to scale with details without consuming a large
number of enums, permitting details that we do not necessarily
want resolution on to propagate to TRON.

Test: boot_reason_test.sh
Bug: 116838876
Bug: 128929506
Change-Id: I3f7b5de22f99195b7ce99672a212174365960b3f