Update patch set 5
Patch Set 5:

This is a bit racy though isn't it?  The first person to call SetTaskProfiles() gets to create the singleton with or without caching that the rest of the process will have to deal with.  It fixes the JNI case now, but what if a new call is added (potentially even in a dependency) that creates the singleton with the cached FDs?

That said, I don't have a great idea; unless maybe you add a function like DropFdCaches() that closes the cached FDs and call that after forking?

Patch-set: 5
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>