Reland^2: "init: run property service in a thread"
It's been a long standing issue that init cannot respond to property
set messages when it is running a builtin command.  This is
particularly problematic when the commands involve IPC to vold or
other daemons, as it prevents them from being able to set properties.

This change has init run property service in a thread, which
eliminates the above issue.

This change may also serve as a starting block to running property
service in an entirely different process to better isolate init from
handling property requests.

Reland: Drop the logic for sending a 'LoadPersistentProperties'
  message and drop the queue for property_service to respond to
  control messages, preferring instead for init to directly respond.
Reland^2: Fix race condition with releasing an fd from unique_fd and
  fdsan.

Test: CF boots, walleye boots, properties are set appropriately
Change-Id: I22c9c33a766a83f2c38613da2e35505d919f29f6