init: simplify async restorecon
In the future, property service may run in its own thread, which means
that PropertyChildReap() needs to be refactored to not run as part of
the init signal handler.

The new method spawns a new thread that handles the queue of paths
that require restorecon. It then communicates back to init via
android::base::SetProperty(). Init distinguishes the thread from other
callers of SetProperty() by checking the pid in the credentials for
the socket connection, thus avoiding dependencies on the rest of
init.

The new method also drops the genericness, since restorecon is the
only function that we should ever need to run asynchronously

Test: async restorecon works, including with queued requests
Change-Id: I2ca00459969e77b1820776dac23d0a0d974e330b