Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> From what I can tell, it seems to be correct to add the runtime APEX to search paths because it is not a sufficient but rather a necessary condition for a library to be exported for it to appear in search paths. In order for a library to be exported from a namespace, it also needs to be present in shared_libs.

No. 

namespace.default.search.paths  = /system/${LIB}
namespace.default.search.paths += /apex/com.android.runtime/${LIB}

This allows "every" libs in the apex directory be loaded into the default namespace. This is not what we want. We want libs in the apex directory to be loaded in the 'runtime' namespace which isolated from the default namespace.

Patch-set: 2