Support booting a chained boot-debug.img
boot-debug.img is introduced to allow 'adb root' if the device is
unlocked, and it cannot be release signed. If /boot partition is chained
in AVB signing and boot-debug.img is used, avb_slot_verify() in userspace
will return AVB_SLOT_VERIFY_RESULT_ERROR_PUBLIC_KEY_REJECTED and refuses
to boot. This CL treats the public key rejection as non-fatal for
chained vbmeta to continue booting, if the device is unlocked.

Bug: 129508966
Test: can root with user load which /boot chained in AVB signing
Change-Id: Idfa8caffbb96f33702b1749afd2e2a59616ddba7