unique_fd - do not close fds < 0
This allows storing various error codes in unique_fds without close(-N)
being automatically called from the destructor.

For example in adb, sometimes the error value -2 is stored instead of -1.

In netd we have a convention to return -errno on error (like native linux
syscalls, as opposed to libc wrappers which return -1 and set errno)
and it requires needless contortions to prevent close(-Exxx) from being called.

The kernel requires a ton of storage per file descriptor handle,
at bare minimum there is an array[max_fds] of null pointers to the
actual file objects.

Since we no longer allow negative fds, we limit our selves to 2 billion
file descriptors per process, but this requires 2 billion pointers...
On a 32-bit system, that's 8GB of ram, and just outright impossible.
On a 64-bit system, that's 16GB of ram... so impossible on all current
android devices, and of course even if you actually had more than 16GB
of ram, you'd have to be insane to want to use it on 2 billion fds.

Test: booted a cuttlefish instance, TreeHugger will do more
Signed-off-by: Maciej Żenczykowski <maze@google.com>
Change-Id: I05dea878f68dc8db09172ce88101f95083fbcdd4