Wait for keymaster.status be ready before vold install key
After enabling FBE the system cannot boot into UI correctly, it will
enter factory reset mode randomly.

The root cause is that the keymaster treble service is not ready when
the VOLD try to use it, the orignal design is that all services will be started
in a forked thread and the main thread will continue, while the time interval
between the action of "installKey" and "class_start early_hal" is very small,
sometimes, the "installKey" could not find the keymaster instance due to it's
not ready yet before enumerated by vold.

The solution is to add a new prop "keymaster.status", and wait its value be
ready before the "installKey".

Change-Id: Ic768936e7327819a10cad505bedb3013ddb3a2f4