libprocessgroup_setup: exclude controllers that failed to mount in cgrouprc
Controllers listed in cgroups.json file might fail to mount if kernel is
not configured to support them. In such a case there is not point of
recording them as valid controllers. Skip controllers that failed to mount
when generating cgroups.rc file. This results in cpusets_enabled() and
schedboost_enabled() functions returning "false" when corresponding
controller failed to mount and therefore reports the actual state of the
controller.

Bug: 124080437
Test: libcutils_test
Change-Id: I770cc39fe50465146e3205aacf77dc3c56923c5d
Signed-off-by: Suren Baghdasaryan <surenb@google.com>