init builtins: Add --force flag to restorecon
Since the SELinux libs in platform/external/selinux/ support
SELINUX_ANDROID_RESTORECON_FORCE, the option should be available in init
as well.

Reasoning:
When running restorecon, an xattr named "security.restorecon_last"
containing the hash sums of the file_contexts file(s) is written to the
folder on which restorecon was invoked.
On the next restorecon run, init will skip that folder since the xattr
still matches the hash sums.

That optimization is fine only as long as we can guarantee that no
program or service ever changes labels of files nested inside a folder
which has the "security.restorecon_last" xattr.

For those edge cases, currently one would have to resort to init
trickery such as:
    # Remove "security.restorecon_last" xattr from /persist/
    # to force restorecon
    exec u:r:vendor_init:s0 root root -- /vendor/bin/toybox_vendor \
      setfattr -x security.restorecon_last /persist/
    restorecon_recursive /persist/

The performance penalty incurred when forcing a restorecon should be
neglible for e.g. /persist partitions or other folders which typically
only hold about 50 files.

Signed-off-by: Felix Elsner <google@ix5.org>
Signed-off-by: MarijnS95 <marijns95@gmail.com>
Change-Id: I1258a6a9974843915602679fa95c635b0c53af03