Explain the usage of firstMarkedBit
The function firstMarkedBit counts from the left. It returns the
position of the leftmost bit, but the position is counted from the left.
That's not natural, because it would produce different results for 32 bit and 64
bit values.

Rather than 'fix' the function to produce the expected result, document
the existing behaviour and add tests.

Bug: 131927078
Test: atest libutils_test
Change-Id: I18c2c5fa5ef80c677da621d4de9667fa0f8d4368