Update patch set 1
Patch Set 1:

Agreed that we need to clean this up a bit.  enh@ would love us to reboot hard if /data is not mounted.

What about the nofail fs_mgr option though?  Presumably we want to respect that instead of aborting.  Also, for the early returns, there's always the chance that the partition is one of a few 'alternatives' so not being able to set it up would still be valid.

(We probably need to refactor this whole thing anyway; it's been on my list; not high enough though)

Patch-set: 1
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>