Update patch set 3
Patch Set 3:

Hey Nick, thank you for taking the time to comment.

What Marijn and me were seeing when we decided to write this CL were cases where our own Android Pie builds inexplicably would not "fix" wrong SELinux labels on the /persist partition. Those wrong labels had been set by the stock firmware flashed in between, which apparently had updated the file labels but not updated the xattr.

Now of course you could argue that we should rather be fixing whatever mess that caused this, but that's not always feasible when dealing - as hobbyists - with closed-source vendor firmware.

I think the restorecon_recursive command should be a guarantee that a folder is in a certain labeled condition. The xattr magic is primarily a speed optimization, and badly needed for e.g. /data with thousands of files. But it should be bypassable in rare cases. Regular vendors with full control over all firmware won't need and won't use this.

If you feel this is contributing to bloating the init binary and thus extending the attack surface, I absolutely understand.

Patch-set: 3
Reviewer: Gerrit User 1421169 <1421169@85c56323-6fa9-3386-8a01-6480fb634889>