Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Should we automatically enable native bridge for transitive dependencies of some root set?  Can we determine the root set automatically from the list of NDK libraries?

I will take a look at flagging dependencies automatically. I initially wanted it to be explicit to see if it drags to many modules in native_bridge, seems not to be the case so far and if it does there is not much we can do about that.

As for root when bridged library replaces NDK library the ndk library does not need to have native_bridge_supported. The list of NDK libraries which are supported 'as is' is rather small.

In the meantime, can you please approve this change?

Patch-set: 1
Reviewer: Gerrit User 1041673 <1041673@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0