init: make LOG(FATAL) reboot target configurable
Currently, LOG(FATAL) reboots to fastboot if init encounters an issue,
but this may be not desirable in all cases, especially the case of
critical services crashing.  Therefore this change adds the ability
for vendors to customize the reboot target via the
ro.build.init_log_fatal_target property.

Note that due to the fact properties are not loaded until 2nd stage
init has made some progress, it will default to fastboot until
properties can be loaded.

This doesn't apply to the reboot to fastboot signal handlers used for
userdebug/eng builds either.

Bug: 121006328
Test: device reboots to configurable target with LOG(FATAL)
Change-Id: I16ea9e32e2fee08dece3d33b697d7a08191d607b