fs_mgr: overlay: do not allow sub-mounts to be overridden
Prevent sub-mounts from being candidates for overlayfs overrides.
We move them temporarily away, then move them back after overlayfs
mount.

There is a race condition that can not be solved where the content
of the sub-mount is temporarily unavailable.  This operation occurs
in first stage init (no problem) and during the first adb remount or
adb disable-verity that causes the backing storage to be setup.

From the manpage: Note also that moving a mount residing under a
shared mount is invalid and unsupported.

Test: adb-remount-test.sh and manual device configs
Bug: 130131892
Change-Id: I8782d09b7b389d899aed07b3e6c528280af2d6c8