Update patch set 1
Patch Set 1:

I think we can refactor even more into these functions and it'll help reviewing the other CL.  

I'd create a struct for ProcessAttributes that contains rlimits_, capabilities_, uid_, gid_, supp_gids_, and priority_, and pass that into a refactored SetProcessAttributes().  You'll have to do a bit of deserialization from the proto, but it'll keep the core logic clean.  Additionally, ioprio_class_ and ioprio_pri_ could be added and moved into that same function.  The call to android_set_ioprio() logically belongs in SetProcessAttributes(); I'm not sure why it's not there now.

The same with the namespaces, namespace_flags_, namespaces_to_enter_, and maybe a few more, should allow you to replace all of the namespace handling code with a single function call.

I'd also move ZapStdio() and OpenConsole() into this file instead of util.cpp (and probably better to do so in this CL, to reduce the changes to review in the other).

Patch-set: 1
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>