Update patch set 5
Patch Set 5:

> Patch Set 5:
> > This fixes two corner cases: eng builds that require phone services (firmware_mnt)
> 
> Is there a difference between eng and userdebug?

One has verity enabled, the other does not.

With verity enabled on userdebug, the adb remount -R or adb disable-verity does not perform any mounts, but the the pre-load of the backign storage, first stage init takes it over as ordered and expected.

One eng builds, the act of adb remount performed the remount immediately b/c verity is disabled. Coming in late, and not at first stage init, things are out of place, and most notably the sub-mount becomes invisible (hence this CL).

>  I know the bug was filed because of userdebug builds; does the previous patch that you already submitted fix them?

Yes, unless they have the other corner condition (sub-mount in first stage init)

> 
> > device configurations that mount sub-mounts in first stage init
> 
> Alternatively, instead of calling overlay mount all, could we mount overlayfs one by one in first stage init while the mounts are being performed?  This would have the same net effect with the least side effects, right?

Yes. The next level of KISS is to move fs_mgr_overlayfs into init (if adb remount, adb disable/enable-verity did not also need to depend on it).

The original design of overlayfs was to be done once, isolated/orthogonal from init, and performed at the end to keep the impact on init low, boot time impact low. This layered approach has cost us with some complexity.

> We may want to allow nesting overlayfs in this case though.  Since right now, the parent overlayfs overlays the submounts.

That is the problem, the parent overlayfs does _not_ always nest the submounts, for some reason it also can _hide_ them, hence this workaround (!). I decided to provide the least common denominator since we did not support adb remount, and thus not overriding content, of the submounts in the first place.

> If we either made this change or did the above, that wouldn't be the case.

> 
> I believe apps run in a mount namespace, https://cs.corp.google.com/android/frameworks/base/core/jni/com_android_internal_os_Zygote.cpp?q=unshare+f:frameworks/base&sq=package:%5Eandroid$&g=0&l=619, so getting public vs private may be important if these are paths apps may access.  Init allows launching a service in a mount namespace too.
> 
> Maybe the craziest idea, if we did allow nesting overlayfs mounts,

There is a limit of 6 or 8 AFAIK as to how deep one can nest overlayfs. PRobably OK for Android's use case.

> this problem may be easier.  We would start with the most specific mount points, and create the overlayfs mounts, but make the mount point a temp directory, e.g. `mount -t overlayfs -o lowerdir=<lower>,upperdir=<real upper dir> overlayfs /dev/tmp-XXXX.  Then move these in place from the least specific to the most specific.

IDK if overlayfs will allow this on all kernels ...

> Lastly, does the dirfd example even work with overlayfs in general?  What happens if a process has a dirfd, then overlayfs is mounted, then the process calls openat() with that dirfd?  Is overlayfs good enough to intercept that openat() call?  I'm guessing not actually, so maybe not the biggest thing to worry about; this is only for developers and only if the device isn't rebooted.

dirfd will be referencing the underlying filesystem. So no.

Overlayfs in general is _different_ in each kernel release, so these subtle behavioral differences can not be answered with just a yes or a no sometimes :-( My testing of overlayfs has suffered with scale of testing (have to test on _all_ devices and _all_ kernels, can not use adb remount to test my changes, so a _full_ flash of the device, etc etc etc). One change takes me a whole day to test sometimes ...

Patch-set: 5
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0