Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Why do you need to set these one at a time?
> 
> Why are you running logd as root in the first place?
> 
> I think there're a handful of changes that we can make here, but I don't think those are it.

Hi Tom,

I'm helping to test the code. 

Since we call setuid in the code, we need CAP_SETUID to do it. With current implementation it works even we don't have CAP_SETUID capability yet. I works magically because it is run with user logd.

So the fix is, either remove setuid, or add CAP_SETUID capability, both should be OK. But if we remove setuid, then it can only start by user logd, otherwise we can not drop the privilege.

With the fix in this patch, no matter which user it is run by, it will always work well.

Thanks,
Lianwei

Patch-set: 1
Reviewer: Gerrit User 1001591 <1001591@85c56323-6fa9-3386-8a01-6480fb634889>