Update patch set 2
Patch Set 2:

Let's call it something else then?  Returning a std::string when we don't need to is kind of unfortunate.  Though I guess good enough given the intentions of this CL.

I guess it only matters if we start using string_view as a Split return value (Which admittedly, we're still pretty against), but we'd end up with something like:

for (string_view line : Split(sv_file, '\n')) {
  Trim(&line);
  for (string_view token : Split(line, ' ')) {
    // Do something with token
  }
}

If we'd either need a Trim() that takes string_view* or one that return string_view, and the former is safer than the latter.

Patch-set: 2
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>