Update patch set 2
Patch Set 2:

(2 comments)

> Patch Set 2:
> 
> Let's call it something else then?  Returning a std::string when we don't need to is kind of unfortunate.  Though I guess good enough given the intentions of this CL.
> 
> I guess it only matters if we start using string_view as a Split return value (Which admittedly, we're still pretty against), but we'd end up with something like:
> 
> for (string_view line : Split(sv_file, '\n')) {
>   Trim(&line);
>   for (string_view token : Split(line, ' ')) {
>     // Do something with token
>   }
> }
> 
> If we'd either need a Trim() that takes string_view* or one that return string_view, and the former is safer than the latter.

hmm... this might be why we ground to a halt and did nothing last time :-/

but i still think i'd argue (and it sounds like you agree) that if/when we really care about this specific "iterate trimmed lines without allocation" case, we need a bigger hammer anyway. so match the existing API for now (which was all neel wanted) and wait until we're forced to have a better idea?

Patch-set: 2
Reviewer: Gerrit User 1003224 <1003224@85c56323-6fa9-3386-8a01-6480fb634889>