Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> (1 comment)
> 
> +lavers who implemented something similar here already.
> 
> I don't really know what to say here.
> 
> This is definitely wrong when we think about how the treble model was designed.  There's a product partition that should be providing these values and that partition should be different between SKUs.  Having a set of 'product' properties on the ODM partition and switching which ones we load by SKU is wrong.
> 
> But if we really still need to do this, then this change itself isn't bad.  It's just unfortunate.

I understand the need. OEMs often use the same partitions across multiple SKUs to reduce the number of builds that they should maintain. In some verticals, especially like auto, the number of SKUs are so many and the granularity of the customization is very fine. In that case, partition-based customization doesn't work well. I don't think that is against the spirit of Treble.

However, what I don't understand with this change is the reason why the sku-specific build props are in the odm partition, not in product, vendor, etc.

Patch-set: 1