Update patch set 4
Patch Set 4:

(1 comment)

Look good to me.  One idea inline, but just an idea.

More context for enh@ for why this belongs in android-base: this is used by a dependency of fs_mgr, so we can't put it easily in fs_mgr itself, and it's used by all of apexd, vold, fs_mgr, and init, so we have at least 4 users.

Patch-set: 4
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>