Update patch set 1
Patch Set 1:

I'm still reviewing this, but I'm not too sure that we're going to take this CL as is.  I have a lot of concerns with using the realtime time for the region lock and while this CL may make this specific situation better, I think there are still a lot of dangers in the overall design here.

Patch-set: 1
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>