Add retrieval of package name for host-side APK
Fastdeploy requires that knowledge of the APK package name on the host
side. The code to parse the compiled xml format lives in frameworks/base
which the PDK cannot depend on. This change copies necessesary code out
of frameworks/base to parse the binary AndroidManifest.xml and retrieve
the package name.

Bug: 135206892
Test: Enable fastdeploy, install APK using --fastdeploy flag, log
      parsed APK package name

Change-Id: Ia55d3c5a1f2fdb747569c32f6a53ed23a6c32c91