Update patch set 3
Patch Set 3:

I take it that we're not accessing these files through a HAL, but rather there's an ABI formed with them that's Treble compliant and vendor daemons put them here?

If so, this is fine.

If not, then we should put this line in the vendor files, particularly with the init.rc file associated with the HAL that generates these.  Remember that you can have on post-data-fs lines in HAL init.rc files too.

Patch-set: 3
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>