libdm: Improve the reliability of dm device paths.
This fixes a race condition where WaitForFile() after
DmGetDevicePathByName appears to succeed, but a subsequent operation on
the path fails. This can happen when CreateDevice() is called
immediately after a call to DeleteDevice (from any process), and the
path is re-used, enqueuing udev events to remove and re-add the block
device.

The fix for this is to introduce a new variant of CreateDevice() that
has a timeout parameter. When the timeout is positive, CreateDevice()
will wait for both of the following conditions to be true:
 1. The /dev/block/dm-N path exists.
 2. The dm/uuid node in sysfs matches a random uuid assigned when we
    created the device.

This ensures that callers of CreateDevice() can use the device path
knowing that no asynchronous removals are pending. Code that uses the
old CreateDevice+WaitForFile pattern will be transitioned to the new
method. This has already been done for code under fs_mgr, and for code
that added a timeout to DeleteDevice as a workaround (for example,
fastbootd).

Note that it is safe to ignore the timeout, or to use the "unsafe"
CreateDevice, if the caller ensures the path by other means. For example
first-stage init has no device removal, and regenerates uevents until
it has acquired all the paths it needs.

Finally, since libdm now inspects sysfs unconditionally, libdm consumers
need r_dir_file perms for sysfs_dm in their sepolicy.

Bug: 135771280
Test: libdm_test
      device boots
      gsid works
Change-Id: If5a7383ea38f32a7fbbcf24842dce6a668050a70